DEMOS-32: moved tests to own file, added debounce, added localstorage persistance
